+++ CHANGELOG.md
@@ -5,11 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.2] - 2026-08-01
 
 * Gem metadata
 * The published gem no longer ships non-production code
-
 
 ## [1.1.1] - 2026-07-31